Corrections following review
