+++ homeassistant/components/flipr/strings.json
@@ -11,8 +11,8 @@
         }
       },
       "flipr_id": {
-        "title": "Enter or choose your Flipr",
-        "description": "Flipr ID not determined. Enter or choose your Flipr ID.",
+        "title": "Choose your Flipr",
+        "description": "Choose your Flipr ID in the list",
         "data": {
           "flipr_id": "Flipr ID"
         }
@@ -21,7 +21,8 @@
     "error": {
       "cannot_connect": "[%key:common::config_flow::error::cannot_connect%]",
       "invalid_auth": "[%key:common::config_flow::error::invalid_auth%]",
-      "unknown": "[%key:common::config_flow::error::unknown%]"
+      "unknown": "[%key:common::config_flow::error::unknown%]",
+      "no_flipr_id_found": "No flipr id associated to your account for now. You should verify it is working with the Flipr's mobile app first."
     },
     "abort": {
       "already_configured": "[%key:common::config_flow::abort::already_configured_device%]"
